requirements.txt
fastapi
transformers
datasets 
accelerate
uvicorn
python-multipart
pillow
torchvision
pandas